requirements.txt
ollama>=0.3.0
openai>=1.0.0
pyyaml>=6.0
